safari-600.7-branch/LayoutTests/http/tests/security/video-cross-origin-accessfailure-expected.txt

EVENT(error)
END OF TEST
 
